Improved pathing algorithm + bug fixes

+++ src/hooks/useFirestore.js
@@ -1,4 +1,4 @@
-import { useState, useEffect } from 'react';
+import { useState, useEffect, useRef } from 'react';
 import { projectFirestore, projectStorage } from '../firebase/config';
 import Cookies from 'universal-cookie';
 
@@ -7,6 +7,7 @@ const cookies = new Cookies();
 const useFirestore = (collection) => {
     
     const [doc, setDoc] = useState(null);
+    const isDeleting = useRef(false);
 
     useEffect(() => {
 
@@ -18,14 +19,16 @@ const useFirestore = (collection) => {
                     setDoc({ ...docRef.data(), id: docRef.id });
                 }
                 // Delete expired images
-                if(docRef.data().createdAt != null) {
-                    if(docRef.data().createdAt.seconds + 12*3600 < Date.now() / 1000) {
+                if(!isDeleting.current && docRef.data().createdAt != null) {
+                    if(docRef.data().createdAt.seconds + 2*3600 < Date.now() / 1000) {
+                        isDeleting.current = true;
                         projectFirestore.collection(collection).where('__name__', '==', docRef.id).get().then(function(querySnapshot) {
                             querySnapshot.forEach(function(doc) {
                                 doc.ref.delete();
                             });
                         });
                         projectStorage.refFromURL(docRef.data().url).delete();
+                        setTimeout(() => isDeleting.current = false, 100);
                     }
                 }
             });
